% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{add_data_layer}
\alias{add_data_layer}
\title{Add data layer to plot based on plot type}
\usage{
add_data_layer(p, panel, color_by)
}
\arguments{
\item{p}{ggplot object}

\item{panel}{Panel configuration}

\item{color_by}{Column to map to color}
}
\value{
Updated ggplot object
}
\description{
Add data layer to plot based on plot type
}
\keyword{internal}
