% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_highlight_overlay}
\alias{add_highlight_overlay}
\title{Add highlight overlay to a panel}
\usage{
add_highlight_overlay(p, highlight)
}
\arguments{
\item{p}{ggplot object}

\item{highlight}{Data frame with start/end}
}
\value{
Updated ggplot object
}
\description{
Add highlight overlay to a panel
}
\keyword{internal}
