% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{add_vlines_to_plot}
\alias{add_vlines_to_plot}
\title{Add vertical lines to a plot}
\usage{
add_vlines_to_plot(p, vlines_data, x_limits)
}
\arguments{
\item{p}{ggplot object}

\item{vlines_data}{List of vline specifications}

\item{x_limits}{X-axis limits}
}
\value{
Updated ggplot object
}
\description{
Add vertical lines to a plot
}
\keyword{internal}
