% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_x_axis}
\alias{add_x_axis}
\title{Add X-axis to bottom panel}
\usage{
add_x_axis(p, region)
}
\arguments{
\item{p}{ggplot object}

\item{region}{Viewing region}
}
\value{
Updated ggplot object
}
\description{
Add X-axis to bottom panel
}
\keyword{internal}
