% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{apply_panel_scales}
\alias{apply_panel_scales}
\title{Apply scales to panel}
\usage{
apply_panel_scales(p, panel, x_limits)
}
\arguments{
\item{p}{ggplot object}

\item{panel}{Panel configuration}

\item{x_limits}{X-axis limits}
}
\value{
Updated ggplot object
}
\description{
Apply scales to panel
}
\keyword{internal}
