% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{apply_panel_theme}
\alias{apply_panel_theme}
\title{Apply theme to panel}
\usage{
apply_panel_theme(p, panel)
}
\arguments{
\item{p}{ggplot object}

\item{panel}{Panel configuration}
}
\value{
Updated ggplot object
}
\description{
Apply theme to panel
}
\keyword{internal}
