% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{apply_single_transform}
\alias{apply_single_transform}
\title{Apply a single transform to a vector}
\usage{
apply_single_transform(x, pos = NULL, type, params)
}
\arguments{
\item{x}{Numeric vector to transform}

\item{pos}{Position vector (for position-aware transforms)}

\item{type}{Transform type}

\item{params}{Transform parameters}
}
\value{
Transformed vector
}
\description{
Apply a single transform to a vector
}
\keyword{internal}
