% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.R
\name{browser_add_transform}
\alias{browser_add_transform}
\title{Add a transform to a panel}
\usage{
browser_add_transform(browser, panel_name, type, ...)
}
\arguments{
\item{browser}{Browser object}

\item{panel_name}{Name of panel to add transform to}

\item{type}{Transform type: "smooth", "log2", "log10", "sqrt", "zscore", "minmax", "clip", "expr"}

\item{...}{Transform-specific parameters}
}
\value{
Updated browser object
}
\description{
Add a transform to a panel
}
