% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{browser_clear_cache}
\alias{browser_clear_cache}
\title{Clear the entire cache (memory and optionally disk)}
\usage{
browser_clear_cache(disk = TRUE)
}
\arguments{
\item{disk}{Also clear disk cache (default: TRUE)}
}
\description{
Clear the entire cache (memory and optionally disk)
}
