% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{browser_clear_highlight}
\alias{browser_clear_highlight}
\title{Clear highlight region}
\usage{
browser_clear_highlight(browser)
}
\arguments{
\item{browser}{Browser object}
}
\value{
Updated browser object
}
\description{
Clear highlight region
}
\keyword{internal}
