% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{browser_create_config}
\alias{browser_create_config}
\title{Create a new browser configuration programmatically}
\usage{
browser_create_config(misha_root = NULL, title = "Genome Browser")
}
\arguments{
\item{misha_root}{Path to misha database root}

\item{title}{Browser title}
}
\value{
Configuration list
}
\description{
Create a new browser configuration programmatically
}
