% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.R
\name{browser_get_region}
\alias{browser_get_region}
\title{Get current region}
\usage{
browser_get_region(browser)
}
\arguments{
\item{browser}{Browser object}
}
\value{
Current region as data frame
}
\description{
Get current region
}
