% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{browser_goto_gene}
\alias{browser_goto_gene}
\title{Go to gene location}
\usage{
browser_goto_gene(browser, gene, span = NULL)
}
\arguments{
\item{browser}{Browser object}

\item{gene}{Gene name}

\item{span}{Span around gene center}
}
\value{
Updated browser object
}
\description{
Go to gene location
}
\keyword{internal}
