% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{browser_move_right}
\alias{browser_move_right}
\title{Move view right}
\usage{
browser_move_right(browser, fraction = .DEFAULT_MOVE_FRACTION)
}
\arguments{
\item{browser}{Browser object}

\item{fraction}{Fraction of view width to move (default 0.5)}
}
\value{
Updated browser object
}
\description{
Move view right
}
