% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.R
\name{browser_save}
\alias{browser_save}
\title{Save browser configuration to file}
\usage{
browser_save(browser, file)
}
\arguments{
\item{browser}{Browser object}

\item{file}{Output file path}
}
\value{
Browser object (invisibly)
}
\description{
Save browser configuration to file
}
