% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-server.R
\name{browser_server}
\alias{browser_server}
\title{Create browser server function}
\usage{
browser_server(browser)
}
\arguments{
\item{browser}{Browser object}
}
\value{
Shiny server function
}
\description{
Create browser server function
}
\keyword{internal}
