% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{browser_set_highlight}
\alias{browser_set_highlight}
\title{Add highlight region to all panels}
\usage{
browser_set_highlight(browser, start, end)
}
\arguments{
\item{browser}{Browser object}

\item{start}{Highlight start}

\item{end}{Highlight end}
}
\value{
Updated browser object
}
\description{
Add highlight region to all panels
}
\keyword{internal}
