% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.R
\name{browser_set_region}
\alias{browser_set_region}
\title{Set current region}
\usage{
browser_set_region(browser, region)
}
\arguments{
\item{browser}{Browser object}

\item{region}{Data frame with chrom, start, end or coordinate string}
}
\value{
Updated browser object
}
\description{
Set current region
}
