% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.R
\name{browser_set_tracks}
\alias{browser_set_tracks}
\title{Set tracks for a panel}
\usage{
browser_set_tracks(browser, panel_name, tracks)
}
\arguments{
\item{browser}{Browser object}

\item{panel_name}{Name of panel}

\item{tracks}{New track list}
}
\value{
Updated browser object
}
\description{
Set tracks for a panel
}
