% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.R
\name{browser_set_ylim}
\alias{browser_set_ylim}
\title{Set Y-axis limits for a panel}
\usage{
browser_set_ylim(browser, panel_name, ylim)
}
\arguments{
\item{browser}{Browser object}

\item{panel_name}{Name of panel}

\item{ylim}{Y-axis limits as c(min, max)}
}
\value{
Updated browser object
}
\description{
Set Y-axis limits for a panel
}
