% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ui.R
\name{browser_ui}
\alias{browser_ui}
\title{Create browser UI}
\usage{
browser_ui(browser)
}
\arguments{
\item{browser}{Browser object (for configuration)}
}
\value{
Shiny UI definition
}
\description{
Create browser UI
}
\keyword{internal}
