% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{browser_zoom_in}
\alias{browser_zoom_in}
\title{Zoom in on current region}
\usage{
browser_zoom_in(browser, factor = .DEFAULT_ZOOM_FACTOR)
}
\arguments{
\item{browser}{Browser object}

\item{factor}{Zoom factor (2 = double resolution, default)}
}
\value{
Updated browser object
}
\description{
Zoom in on current region
}
