% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{browser_zoom_out}
\alias{browser_zoom_out}
\title{Zoom out from current region}
\usage{
browser_zoom_out(browser, factor = .DEFAULT_ZOOM_FACTOR)
}
\arguments{
\item{browser}{Browser object}

\item{factor}{Zoom factor (2 = half resolution, default)}
}
\value{
Updated browser object
}
\description{
Zoom out from current region
}
