% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-config-editor.R
\name{build_config_from_inputs}
\alias{build_config_from_inputs}
\title{Build configuration from form inputs}
\usage{
build_config_from_inputs(input, vtracks, panels, colors, vlines, base_cfg)
}
\description{
Build configuration from form inputs
}
\keyword{internal}
