% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_exists}
\alias{cache_exists}
\title{Check if key exists in cache (memory or disk)}
\usage{
cache_exists(key)
}
\arguments{
\item{key}{Cache key}
}
\value{
TRUE if key exists
}
\description{
Check if key exists in cache (memory or disk)
}
\keyword{internal}
