% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_set}
\alias{cache_set}
\title{Set value in cache (memory and optionally disk)}
\usage{
cache_set(key, value)
}
\arguments{
\item{key}{Cache key}

\item{value}{Value to cache}
}
\value{
Invisibly returns the value
}
\description{
Set value in cache (memory and optionally disk)
}
\keyword{internal}
