% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_iterator}
\alias{calc_iterator}
\title{Calculate dynamic iterator based on span}
\usage{
calc_iterator(span, base_iter = 32, target_points = 4000)
}
\arguments{
\item{span}{Viewing span in bp}

\item{base_iter}{Base iterator from config}

\item{target_points}{Target number of points (default 4000)}
}
\value{
Calculated iterator value
}
\description{
Ensures approximately target_points data points regardless of view span.
}
\keyword{internal}
