% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-config-editor.R
\name{clean_config_for_export}
\alias{clean_config_for_export}
\title{Clean configuration for YAML export}
\usage{
clean_config_for_export(cfg)
}
\description{
Clean configuration for YAML export
}
\keyword{internal}
