% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{cleanup_temp_vtracks}
\alias{cleanup_temp_vtracks}
\title{Clean up temporary vtracks}
\usage{
cleanup_temp_vtracks(vtracks)
}
\arguments{
\item{vtracks}{Character vector of vtrack names to remove}
}
\description{
Clean up temporary vtracks
}
\keyword{internal}
