% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-config-editor-server.R
\name{config_editor_server}
\alias{config_editor_server}
\title{Config Editor Module Server}
\usage{
config_editor_server(id, browser_rv, original_config)
}
\arguments{
\item{id}{Module namespace ID}

\item{browser_rv}{Reactive value containing browser object}

\item{original_config}{Original configuration for reset}
}
\value{
NULL (module server)
}
\description{
Config Editor Module Server
}
\keyword{internal}
