% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-config-editor.R
\name{config_editor_ui}
\alias{config_editor_ui}
\title{Config Editor Module UI}
\usage{
config_editor_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
Shiny UI element
}
\description{
Config Editor Module UI
}
\keyword{internal}
