% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{convert_plot_type}
\alias{convert_plot_type}
\title{Convert misha.vis plot_type to misha.browser format}
\usage{
convert_plot_type(vis_type)
}
\arguments{
\item{vis_type}{misha.vis plot type}
}
\value{
misha.browser plot type
}
\description{
Convert misha.vis plot_type to misha.browser format
}
\keyword{internal}
