% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{convert_vis_element}
\alias{convert_vis_element}
\title{Convert a single misha.vis element}
\usage{
convert_vis_element(elem)
}
\arguments{
\item{elem}{misha.vis element configuration}
}
\value{
List with panel configuration
}
\description{
Convert a single misha.vis element
}
\keyword{internal}
