% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-config-editor.R
\name{create_config_modal}
\alias{create_config_modal}
\title{Create the config editor modal}
\usage{
create_config_modal(ns, cfg)
}
\arguments{
\item{ns}{Namespace function}

\item{cfg}{Current configuration}
}
\value{
Modal dialog
}
\description{
Create the config editor modal
}
\keyword{internal}
