% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-config-editor.R
\name{create_panel_form}
\alias{create_panel_form}
\title{Create panel editing form}
\usage{
create_panel_form(ns, panel, index, available_tracks)
}
\description{
Create panel editing form
}
\keyword{internal}
