% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.R
\name{create_vtrack}
\alias{create_vtrack}
\title{Create a virtual track from configuration}
\usage{
create_vtrack(vt, cfg = NULL)
}
\arguments{
\item{vt}{Vtrack configuration}

\item{cfg}{Browser configuration (for resolving paths)}
}
\description{
Create a virtual track from configuration
}
\keyword{internal}
