% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-config-editor.R
\name{expand_colors_with_tracks}
\alias{expand_colors_with_tracks}
\title{Expand colors list with track names for UI display}
\usage{
expand_colors_with_tracks(colors, panels)
}
\description{
Expand colors list with track names for UI display
}
\keyword{internal}
