% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{expand_interval}
\alias{expand_interval}
\title{Expand interval to a specific width}
\usage{
expand_interval(interval, target_width)
}
\arguments{
\item{interval}{Interval data frame}

\item{target_width}{Target width in bp}
}
\value{
Expanded interval
}
\description{
Expand interval to a specific width
}
\keyword{internal}
