% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{expand_region}
\alias{expand_region}
\title{Expand region by a given amount}
\usage{
expand_region(region, expansion)
}
\arguments{
\item{region}{Region data frame}

\item{expansion}{Expansion in bp (total, split evenly)}
}
\value{
Expanded region
}
\description{
Expand region by a given amount
}
\keyword{internal}
