% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_annotation_data}
\alias{extract_annotation_data}
\title{Extract annotation data (genes, exons, TSS)}
\usage{
extract_annotation_data(panel, region)
}
\arguments{
\item{panel}{Annotation panel configuration}

\item{region}{Interval data frame}
}
\value{
List with exons, tss, genes data frames
}
\description{
Extract annotation data (genes, exons, TSS)
}
\keyword{internal}
