% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{extract_intervals_data}
\alias{extract_intervals_data}
\title{Extract intervals for the panel}
\usage{
extract_intervals_data(panel, region)
}
\arguments{
\item{panel}{Panel configuration}

\item{region}{Viewing region}
}
\value{
Data frame with intervals
}
\description{
Extract intervals for the panel
}
\keyword{internal}
