% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_tracks}
\alias{extract_tracks}
\title{Extract tracks from misha}
\usage{
extract_tracks(tracks, region, iterator = .DEFAULT_ITERATOR, colnames = NULL)
}
\arguments{
\item{tracks}{Character vector of track names/expressions}

\item{region}{Interval data frame}

\item{iterator}{Iterator value}

\item{colnames}{Column names for the result (defaults to tracks)}
}
\value{
Data frame with extracted data
}
\description{
Extract tracks from misha
}
\keyword{internal}
