% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_coords}
\alias{format_coords}
\title{Format genomic coordinates for display}
\usage{
format_coords(chrom, start, end)
}
\arguments{
\item{chrom}{Chromosome}

\item{start}{Start position}

\item{end}{End position}
}
\value{
Formatted string like "chr1:1,000,000-2,000,000"
}
\description{
Format genomic coordinates for display
}
\keyword{internal}
