% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-config-validation.R
\name{format_validation_errors}
\alias{format_validation_errors}
\title{Format validation errors for display}
\usage{
format_validation_errors(errors)
}
\arguments{
\item{errors}{List of validation errors}
}
\value{
HTML formatted error list
}
\description{
Format validation errors for display
}
\keyword{internal}
