% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlines.R
\name{get_all_vlines}
\alias{get_all_vlines}
\title{Get all vline data for plotting}
\usage{
get_all_vlines(browser, region)
}
\arguments{
\item{browser}{Browser object}

\item{region}{Current viewing region}
}
\value{
List of vline specifications for plotting
}
\description{
Get all vline data for plotting
}
\keyword{internal}
