% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-config-editor-panels.R
\name{get_available_tracks}
\alias{get_available_tracks}
\title{Get available tracks from vtracks and existing panel tracks}
\usage{
get_available_tracks(vtracks, panels)
}
\description{
Get available tracks from vtracks and existing panel tracks
}
\keyword{internal}
