% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideogram.R
\name{get_chromosome_size}
\alias{get_chromosome_size}
\title{Get chromosome size}
\usage{
get_chromosome_size(chrom)
}
\arguments{
\item{chrom}{Chromosome name}
}
\value{
Chromosome size in bp, or NULL
}
\description{
Get chromosome size
}
\keyword{internal}
