% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ui.R
\name{get_default_tracks}
\alias{get_default_tracks}
\title{Get default selected tracks}
\usage{
get_default_tracks(browser)
}
\arguments{
\item{browser}{Browser object}
}
\value{
Character vector of default track names
}
\description{
Get default selected tracks
}
\keyword{internal}
