% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{get_effective_smooth_window}
\alias{get_effective_smooth_window}
\title{Get smooth window from browser state}
\usage{
get_effective_smooth_window(browser, panel)
}
\arguments{
\item{browser}{Browser object}

\item{panel}{Panel configuration}
}
\value{
Effective smooth window value
}
\description{
Adjusts smooth window from config based on current state.
}
\keyword{internal}
