% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_panel}
\alias{get_panel}
\title{Get panel by name}
\usage{
get_panel(cfg, name)
}
\arguments{
\item{cfg}{Configuration list}

\item{name}{Panel name}
}
\value{
Panel configuration or NULL
}
\description{
Get panel by name
}
\keyword{internal}
