% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{get_panel_colors}
\alias{get_panel_colors}
\title{Get colors for panel}
\usage{
get_panel_colors(panel, data, color_by, config_colors = NULL)
}
\arguments{
\item{panel}{Panel configuration}

\item{data}{Data frame}

\item{color_by}{Column name to color by}
}
\value{
Named vector of colors
}
\description{
Get colors for panel
}
\keyword{internal}
