% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ui.R
\name{get_track_choices}
\alias{get_track_choices}
\title{Get track choices for selection UI}
\usage{
get_track_choices(browser)
}
\arguments{
\item{browser}{Browser object}
}
\value{
Character vector of track names
}
\description{
Get track choices for selection UI
}
\keyword{internal}
