% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlines.R
\name{get_vline_positions}
\alias{get_vline_positions}
\title{Get vline positions for plotting}
\usage{
get_vline_positions(vline, intervals)
}
\arguments{
\item{vline}{Vline configuration}

\item{intervals}{Loaded intervals}
}
\value{
Numeric vector of x-positions
}
\description{
Get vline positions for plotting
}
\keyword{internal}
