% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.R
\name{init_start_region}
\alias{init_start_region}
\title{Initialize start region from configuration}
\usage{
init_start_region(browser)
}
\arguments{
\item{browser}{Browser object}
}
\value{
Updated browser object
}
\description{
Initialize start region from configuration
}
\keyword{internal}
