% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.R
\name{init_vtracks}
\alias{init_vtracks}
\title{Initialize virtual tracks from configuration}
\usage{
init_vtracks(browser)
}
\arguments{
\item{browser}{Browser object}
}
\value{
Updated browser object
}
\description{
Initialize virtual tracks from configuration
}
\keyword{internal}
