% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interval_center}
\alias{interval_center}
\title{Get interval center}
\usage{
interval_center(interval)
}
\arguments{
\item{interval}{Interval data frame with start and end}
}
\value{
Center position (numeric)
}
\description{
Get interval center
}
\keyword{internal}
