% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interval_width}
\alias{interval_width}
\title{Get interval width}
\usage{
interval_width(interval)
}
\arguments{
\item{interval}{Interval data frame with start and end}
}
\value{
Width in bp
}
\description{
Get interval width
}
\keyword{internal}
