% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{intervals_exist}
\alias{intervals_exist}
\title{Check if misha intervals exist}
\usage{
intervals_exist(intervals)
}
\arguments{
\item{intervals}{Intervals name}
}
\value{
TRUE if intervals exist
}
\description{
Check if misha intervals exist
}
\keyword{internal}
