% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideogram.R
\name{load_cytoband_data}
\alias{load_cytoband_data}
\title{Load cytoband data}
\usage{
load_cytoband_data(cytoband_file, cytoband_track, chrom)
}
\arguments{
\item{cytoband_file}{Path to cytoband file}

\item{cytoband_track}{Misha intervals name for cytobands}

\item{chrom}{Chromosome to filter to}
}
\value{
Data frame with cytoband data
}
\description{
Load cytoband data
}
\keyword{internal}
