% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{load_navigator_regions}
\alias{load_navigator_regions}
\title{Load navigator regions}
\usage{
load_navigator_regions(browser)
}
\arguments{
\item{browser}{Browser object}
}
\value{
Data frame with regions and labels
}
\description{
Load navigator regions
}
\keyword{internal}
