% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlines.R
\name{load_vline_current}
\alias{load_vline_current}
\title{Load vlines from current region boundaries}
\usage{
load_vline_current(region)
}
\arguments{
\item{region}{Current viewing region}
}
\value{
Intervals data frame
}
\description{
Load vlines from current region boundaries
}
\keyword{internal}
