% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlines.R
\name{load_vline_file}
\alias{load_vline_file}
\title{Load vlines from file}
\usage{
load_vline_file(vline, data_dir = ".")
}
\arguments{
\item{vline}{Vline configuration}

\item{data_dir}{Data directory}
}
\value{
Intervals data frame
}
\description{
Load vlines from file
}
\keyword{internal}
