% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlines.R
\name{load_vline_inline}
\alias{load_vline_inline}
\title{Load vlines from inline definition}
\usage{
load_vline_inline(vline)
}
\arguments{
\item{vline}{Vline configuration}
}
\value{
Intervals data frame
}
\description{
Load vlines from inline definition
}
\keyword{internal}
