% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlines.R
\name{load_vline_intervals}
\alias{load_vline_intervals}
\title{Load vertical line intervals}
\usage{
load_vline_intervals(vline, region, data_dir = ".")
}
\arguments{
\item{vline}{Vline configuration}

\item{region}{Current viewing region}

\item{data_dir}{Data directory for resolving relative paths}
}
\value{
Data frame with positions to draw lines
}
\description{
Load vertical line intervals
}
\keyword{internal}
