% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlines.R
\name{load_vline_misha}
\alias{load_vline_misha}
\title{Load vlines from misha intervals}
\usage{
load_vline_misha(vline)
}
\arguments{
\item{vline}{Vline configuration}
}
\value{
Intervals data frame
}
\description{
Load vlines from misha intervals
}
\keyword{internal}
