% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lookup_color_value}
\alias{lookup_color_value}
\title{Lookup a color allowing facet suffix fallbacks}
\usage{
lookup_color_value(value, colors)
}
\arguments{
\item{value}{Color key to resolve}

\item{colors}{Named list/vector of colors}
}
\value{
Color string or NULL if not found
}
\description{
Lookup a color allowing facet suffix fallbacks
}
\keyword{internal}
