% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_region}
\alias{make_region}
\title{Create a sanitized region data frame}
\usage{
make_region(chrom, start, end)
}
\arguments{
\item{chrom}{Chromosome}

\item{start}{Start position}

\item{end}{End position}
}
\value{
Sanitized region data frame
}
\description{
Create a sanitized region data frame
}
\keyword{internal}
