% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misha.browser-package.R
\docType{package}
\name{misha.browser-package}
\alias{misha.browser}
\alias{misha.browser-package}
\title{misha.browser: Interactive Genome Browser for Misha Tracks}
\description{
A configurable, interactive genome browser built on ggplot2 and patchwork. Provides YAML-based configuration for visualization of misha genomic tracks with support for custom transformations, faceting, vertical line annotations, and gene models. Includes a Shiny application for interactive exploration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tanaylab.github.io/misha.browser/}
}

}
\author{
\strong{Maintainer}: Aviezer Lifshitz \email{aviezer.lifshitz@weizmann.ac.il}

Other contributors:
\itemize{
  \item Weizmann Institute of Science [copyright holder]
}

}
\keyword{internal}
