% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{normalize_strand}
\alias{normalize_strand}
\title{Normalize strand/direction to 1 (forward) or -1 (reverse)}
\usage{
normalize_strand(x)
}
\arguments{
\item{x}{Strand values: "+", "-", 1, -1, "forward", "reverse", etc.}
}
\value{
Integer vector of 1 or -1, NA for unknown
}
\description{
Normalize strand/direction to 1 (forward) or -1 (reverse)
}
\keyword{internal}
