% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ui.R
\name{prepare_region_choices}
\alias{prepare_region_choices}
\title{Prepare region choices for virtual select}
\usage{
prepare_region_choices(regions, label_field)
}
\arguments{
\item{regions}{Data frame with regions}

\item{label_field}{Column to use for labels}
}
\value{
List suitable for virtualSelectInput
}
\description{
Prepare region choices for virtual select
}
\keyword{internal}
