% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.R
\name{print.browser}
\alias{print.browser}
\title{Print method for browser object}
\usage{
\method{print}{browser}(x, ...)
}
\arguments{
\item{x}{Browser object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for browser object
}
