% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideogram.R
\name{render_ideogram_panel}
\alias{render_ideogram_panel}
\title{Render ideogram panel}
\usage{
render_ideogram_panel(panel, region)
}
\arguments{
\item{panel}{Panel configuration}

\item{region}{Viewing region}
}
\value{
ggplot object
}
\description{
Render ideogram panel
}
\keyword{internal}
