% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{render_intervals_panel}
\alias{render_intervals_panel}
\title{Render intervals panel}
\usage{
render_intervals_panel(panel, region, vlines_data = NULL)
}
\arguments{
\item{panel}{Panel configuration}

\item{region}{Viewing region}

\item{vlines_data}{Vertical lines data for overlay}
}
\value{
ggplot object
}
\description{
Render intervals panel
}
\keyword{internal}
