% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{render_panel}
\alias{render_panel}
\title{Render a single panel}
\usage{
render_panel(browser, panel, region, vlines_data, pre_extracted_data = NULL)
}
\arguments{
\item{browser}{Browser object}

\item{panel}{Panel configuration}

\item{region}{Viewing region}

\item{vlines_data}{Vertical lines data}

\item{pre_extracted_data}{Optional pre-extracted data for data panels}
}
\value{
ggplot object
}
\description{
Render a single panel
}
\keyword{internal}
