% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{reset_vtrack_iterators}
\alias{reset_vtrack_iterators}
\title{Reset vtrack iterators to default}
\usage{
reset_vtrack_iterators(vtracks)
}
\arguments{
\item{vtracks}{Character vector of vtrack names}
}
\description{
Removes custom sshift/eshift from virtual tracks by calling
gvtrack.iterator with NULL values.
}
\keyword{internal}
