% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanitize_interval}
\alias{sanitize_interval}
\title{Sanitize an interval}
\usage{
sanitize_interval(interval)
}
\arguments{
\item{interval}{Data frame with chrom, start, end}
}
\value{
Sanitized interval or NULL if invalid
}
\description{
Ensures interval has valid chrom, start, end and that start < end.
}
\keyword{internal}
