% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{set_interval_width}
\alias{set_interval_width}
\title{Set interval to specific width (centered)}
\usage{
set_interval_width(interval, target_width)
}
\arguments{
\item{interval}{Interval data frame}

\item{target_width}{Target width in bp}
}
\value{
Resized interval
}
\description{
Set interval to specific width (centered)
}
\keyword{internal}
