% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_panel}
\alias{set_panel}
\title{Set panel by name}
\usage{
set_panel(cfg, name, panel)
}
\arguments{
\item{cfg}{Configuration list}

\item{name}{Panel name}

\item{panel}{New panel configuration}
}
\value{
Updated configuration
}
\description{
Set panel by name
}
\keyword{internal}
