% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{track_exists}
\alias{track_exists}
\title{Check if a misha track exists}
\usage{
track_exists(track)
}
\arguments{
\item{track}{Track name}
}
\value{
TRUE if track exists
}
\description{
Check if a misha track exists
}
\keyword{internal}
