% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_clip}
\alias{transform_clip}
\title{Clip values to range}
\usage{
transform_clip(x, params)
}
\arguments{
\item{x}{Numeric vector}

\item{params}{List with: min, max}
}
\value{
Clipped vector
}
\description{
Clip values to range
}
\keyword{internal}
