% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_log10}
\alias{transform_log10}
\title{Log10 transform}
\usage{
transform_log10(x, params)
}
\arguments{
\item{x}{Numeric vector}

\item{params}{List with: offset (default 1)}
}
\value{
Log10-transformed vector
}
\description{
Log10 transform
}
\keyword{internal}
