% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_log2}
\alias{transform_log2}
\title{Log2 transform}
\usage{
transform_log2(x, params)
}
\arguments{
\item{x}{Numeric vector}

\item{params}{List with: offset (default 1)}
}
\value{
Log2-transformed vector
}
\description{
Log2 transform
}
\keyword{internal}
