% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_minmax}
\alias{transform_minmax}
\title{Min-max normalization}
\usage{
transform_minmax(x, params)
}
\arguments{
\item{x}{Numeric vector}

\item{params}{List (unused)}
}
\value{
Normalized vector scaled between 0 and 1
}
\description{
Scales values to the range 0 to 1.
}
\keyword{internal}
