% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_quantile}
\alias{transform_quantile}
\title{Quantile normalization}
\usage{
transform_quantile(x, params)
}
\arguments{
\item{x}{Numeric vector}

\item{params}{List with: probs (quantile probabilities, default c(0.01, 0.99))}
}
\value{
Quantile-clipped vector
}
\description{
Quantile normalization
}
\keyword{internal}
