% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_smooth}
\alias{transform_smooth}
\title{Smooth transform (rolling mean)}
\usage{
transform_smooth(x, params)
}
\arguments{
\item{x}{Numeric vector}

\item{params}{List with: window (required), align ("center", "left", "right")}
}
\value{
Smoothed vector
}
\description{
Smooth transform (rolling mean)
}
\keyword{internal}
