% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_sqrt}
\alias{transform_sqrt}
\title{Square root transform}
\usage{
transform_sqrt(x, params)
}
\arguments{
\item{x}{Numeric vector}

\item{params}{List (unused)}
}
\value{
Square root of vector
}
\description{
Square root transform
}
\keyword{internal}
