% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_zscore}
\alias{transform_zscore}
\title{Z-score normalization}
\usage{
transform_zscore(x, params)
}
\arguments{
\item{x}{Numeric vector}

\item{params}{List (unused)}
}
\value{
Z-score normalized vector
}
\description{
Z-score normalization
}
\keyword{internal}
