% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{update_smooth_window}
\alias{update_smooth_window}
\title{Update smooth transforms in panel with new window}
\usage{
update_smooth_window(transforms, smooth_window)
}
\arguments{
\item{transforms}{List of transforms}

\item{smooth_window}{New smooth window value}
}
\value{
Updated transforms list
}
\description{
Update smooth transforms in panel with new window
}
\keyword{internal}
