% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-config-validation.R
\name{validate_colors}
\alias{validate_colors}
\title{Validate color mappings}
\usage{
validate_colors(colors)
}
\arguments{
\item{colors}{Named list of colors}
}
\value{
List of errors
}
\description{
Validate color mappings
}
\keyword{internal}
