% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-config-validation.R
\name{validate_general_section}
\alias{validate_general_section}
\title{Validate general section}
\usage{
validate_general_section(cfg)
}
\arguments{
\item{cfg}{Full configuration}
}
\value{
List of errors
}
\description{
Validate general section
}
\keyword{internal}
