% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_hline}
\alias{validate_hline}
\title{Validate a horizontal line configuration}
\usage{
validate_hline(hline, index, panel_name = "unknown")
}
\arguments{
\item{hline}{Hline config list}

\item{index}{Hline index for error messages}

\item{panel_name}{Panel name for error messages}
}
\value{
Validated hline config
}
\description{
Validate a horizontal line configuration
}
\keyword{internal}
