% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-config-validation.R
\name{validate_navigator_section}
\alias{validate_navigator_section}
\title{Validate navigator section}
\usage{
validate_navigator_section(nav)
}
\arguments{
\item{nav}{Navigator configuration}
}
\value{
List of errors
}
\description{
Validate navigator section
}
\keyword{internal}
