% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_panel}
\alias{validate_panel}
\title{Validate a panel configuration}
\usage{
validate_panel(panel, index)
}
\arguments{
\item{panel}{Panel config list}

\item{index}{Panel index for error messages}
}
\value{
Validated panel config
}
\description{
Validate a panel configuration
}
\keyword{internal}
