% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-config-validation.R
\name{validate_panel_full}
\alias{validate_panel_full}
\title{Validate a panel with comprehensive checks}
\usage{
validate_panel_full(panel, index, vtrack_names = character(0))
}
\arguments{
\item{panel}{Panel configuration}

\item{index}{Index for error naming}

\item{vtrack_names}{Character vector of defined vtrack names}
}
\value{
List of errors
}
\description{
Validate a panel with comprehensive checks
}
\keyword{internal}
