% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-config-validation.R
\name{validate_plot_section}
\alias{validate_plot_section}
\title{Validate plot section}
\usage{
validate_plot_section(plot)
}
\arguments{
\item{plot}{Plot configuration}
}
\value{
List of errors
}
\description{
Validate plot section
}
\keyword{internal}
