% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_vline}
\alias{validate_vline}
\title{Validate a vertical line configuration}
\usage{
validate_vline(vline, index)
}
\arguments{
\item{vline}{Vline config list}

\item{index}{Vline index for error messages}
}
\value{
Validated vline config
}
\description{
Validate a vertical line configuration
}
\keyword{internal}
