% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-config-validation.R
\name{validate_vline_full}
\alias{validate_vline_full}
\title{Validate a vline configuration}
\usage{
validate_vline_full(vline, index)
}
\arguments{
\item{vline}{Vline configuration}

\item{index}{Index for error naming}
}
\value{
List of errors
}
\description{
Validate a vline configuration
}
\keyword{internal}
