% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_vtrack}
\alias{validate_vtrack}
\title{Validate a virtual track configuration}
\usage{
validate_vtrack(vtrack, index)
}
\arguments{
\item{vtrack}{Vtrack config list}

\item{index}{Vtrack index for error messages}
}
\value{
Validated vtrack config
}
\description{
Validate a virtual track configuration
}
\keyword{internal}
