% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-config-validation.R
\name{validate_vtrack_full}
\alias{validate_vtrack_full}
\title{Validate a virtual track with comprehensive checks}
\usage{
validate_vtrack_full(vtrack, index)
}
\arguments{
\item{vtrack}{Vtrack configuration}

\item{index}{Index for error naming}
}
\value{
List of errors
}
\description{
Validate a virtual track with comprehensive checks
}
\keyword{internal}
