% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ui.R
\name{vlines_ui}
\alias{vlines_ui}
\title{Create vlines toggle UI}
\usage{
vlines_ui(browser)
}
\arguments{
\item{browser}{Browser object}
}
\value{
Shiny UI element
}
\description{
Create vlines toggle UI
}
\keyword{internal}
