% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{add_hlines_to_plot}
\alias{add_hlines_to_plot}
\title{Add horizontal lines to a plot}
\usage{
add_hlines_to_plot(p, panel, data, browser = NULL)
}
\arguments{
\item{p}{ggplot object}

\item{panel}{Panel configuration}

\item{data}{Data frame with values for stat calculations}

\item{browser}{Optional browser object (needed for quantile_global)}
}
\value{
Updated ggplot object
}
\description{
Add horizontal lines to a plot
}
\keyword{internal}
