% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.R
\name{browser_add_panel}
\alias{browser_add_panel}
\title{Add a data panel to the browser}
\usage{
browser_add_panel(
  browser,
  name,
  tracks = NULL,
  grouping = NULL,
  facet_by = NULL,
  transforms = list(),
  plot_type = "line",
  colors = NULL,
  ylim = NULL,
  height = 2,
  ...
)
}
\arguments{
\item{browser}{Browser object}

\item{name}{Panel name}

\item{tracks}{Character vector of track names}

\item{grouping}{List with color_by, pattern, overrides}

\item{facet_by}{Variable to facet by (from grouping pattern)}

\item{transforms}{List of transforms to apply}

\item{plot_type}{Plot type: "line", "area", "point", "histogram"}

\item{colors}{Named vector of colors}

\item{ylim}{Y-axis limits}

\item{height}{Relative height}

\item{...}{Additional panel options}
}
\value{
Updated browser object
}
\description{
Add a data panel to the browser
}
