% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_key}
\alias{cache_key}
\title{Generate cache key}
\usage{
cache_key(...)
}
\arguments{
\item{...}{Objects to hash}
}
\value{
MD5 hash string
}
\description{
Creates an MD5 hash key for caching extracted data.
}
\keyword{internal}
