% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_expr}
\alias{transform_expr}
\title{Custom expression transform}
\usage{
transform_expr(x, pos, params)
}
\arguments{
\item{x}{Numeric vector}

\item{pos}{Position vector}

\item{params}{List with: expr (R expression as string)}
}
\value{
Transformed vector
}
\description{
Custom expression transform
}
\keyword{internal}
